# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "D-Form API Documentation",
        "description": "Dokumentasi API lengkap untuk D-Form - Platform manajemen event dan recruitment DOSCOM",
        "contact": {
            "name": "DOSCOM - Dinus Open Source Community",
            "email": "doscom@dinus.ac.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Development Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/recruitments": {
            "get": {
                "tags": [
                    "Recruitments"
                ],
                "summary": "Get all recruitments",
                "description": "Retrieve a list of all recruitments with optional filters",
                "operationId": "728a2201359a26b73fc0a09fd2b555f5",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status (pending, approved, rejected)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "approved",
                                "rejected"
                            ]
                        }
                    },
                    {
                        "name": "division",
                        "in": "query",
                        "description": "Filter by division",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Pemrograman",
                                "Data",
                                "Jaringan",
                                "Medcrev"
                            ]
                        }
                    },
                    {
                        "name": "semester",
                        "in": "query",
                        "description": "Filter by semester",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "1",
                                "3"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name, NIM, email, or phone",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Recruitment"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Recruitments"
                ],
                "summary": "Create a new recruitment",
                "description": "Submit a new recruitment application",
                "operationId": "1db26bbcac707d5e5bd244d070464ced",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_lengkap",
                                    "nim",
                                    "semester",
                                    "nomor_hp",
                                    "email_pribadi",
                                    "email_mahasiswa",
                                    "divisi_utama"
                                ],
                                "properties": {
                                    "nama_lengkap": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "nim": {
                                        "type": "string",
                                        "example": "A11.2023.12345"
                                    },
                                    "semester": {
                                        "type": "string",
                                        "enum": [
                                            "1",
                                            "3"
                                        ],
                                        "example": "3"
                                    },
                                    "nomor_hp": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "email_pribadi": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "email_mahasiswa": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "111202312345@mhs.dinus.ac.id"
                                    },
                                    "divisi_utama": {
                                        "type": "string",
                                        "enum": [
                                            "Pemrograman",
                                            "Data",
                                            "Jaringan",
                                            "Medcrev"
                                        ],
                                        "example": "Pemrograman"
                                    },
                                    "divisi_tambahan": {
                                        "type": "string",
                                        "enum": [
                                            "Pemrograman",
                                            "Data",
                                            "Jaringan",
                                            "Medcrev"
                                        ],
                                        "example": "Data"
                                    },
                                    "username_instagram": {
                                        "type": "string",
                                        "example": "@johndoe"
                                    },
                                    "portofolio": {
                                        "type": "string",
                                        "example": "https://github.com/johndoe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recruitment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Recruitment created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Recruitment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/api/recruitments/{shortUuid}": {
            "get": {
                "tags": [
                    "Recruitments"
                ],
                "summary": "Get recruitment by short UUID",
                "description": "Retrieve a single recruitment by its short UUID",
                "operationId": "b3a00c9e48701cbbb9d57a3ac0ac21ad",
                "parameters": [
                    {
                        "name": "shortUuid",
                        "in": "path",
                        "description": "Recruitment short UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Recruitment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recruitment not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Recruitments"
                ],
                "summary": "Update recruitment",
                "description": "Update an existing recruitment application",
                "operationId": "6d2307c2a703a4593356bb2385c8f575",
                "parameters": [
                    {
                        "name": "shortUuid",
                        "in": "path",
                        "description": "Recruitment short UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_lengkap": {
                                        "type": "string",
                                        "example": "John Doe Updated"
                                    },
                                    "nomor_hp": {
                                        "type": "string",
                                        "example": "081234567899"
                                    },
                                    "divisi_utama": {
                                        "type": "string",
                                        "enum": [
                                            "Pemrograman",
                                            "Data",
                                            "Jaringan",
                                            "Medcrev"
                                        ]
                                    },
                                    "divisi_tambahan": {
                                        "type": "string",
                                        "enum": [
                                            "Pemrograman",
                                            "Data",
                                            "Jaringan",
                                            "Medcrev"
                                        ]
                                    },
                                    "portofolio": {
                                        "type": "string",
                                        "example": "https://github.com/johndoe-updated"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recruitment updated successfully"
                    },
                    "404": {
                        "description": "Recruitment not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Recruitments"
                ],
                "summary": "Delete recruitment",
                "description": "Soft delete a recruitment application",
                "operationId": "7913eeb3a64e38f204971129e765e009",
                "parameters": [
                    {
                        "name": "shortUuid",
                        "in": "path",
                        "description": "Recruitment short UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recruitment deleted successfully"
                    },
                    "404": {
                        "description": "Recruitment not found"
                    }
                }
            }
        },
        "/api/recruitments/{shortUuid}/review": {
            "patch": {
                "tags": [
                    "Recruitments"
                ],
                "summary": "Review recruitment application",
                "description": "Approve or reject a recruitment application (admin only)",
                "operationId": "7335373749803bc7dffe5ceb4757d53b",
                "parameters": [
                    {
                        "name": "shortUuid",
                        "in": "path",
                        "description": "Recruitment short UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "approved",
                                            "rejected"
                                        ],
                                        "example": "approved"
                                    },
                                    "catatan": {
                                        "type": "string",
                                        "example": "Kandidat yang sangat baik"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recruitment reviewed successfully"
                    },
                    "404": {
                        "description": "Recruitment not found"
                    }
                }
            }
        }
    },
...
}